pyproject.toml
[project]
name = "ict-questions-backend"
version = "0.1.0"
description = "API for ICT Questions"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-camel-case>=1.4.2",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
